backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
motor==3.6.*
pymongo>=4.9,<4.10
pydantic==2.*
pydantic-settings==2.*
python-multipart==0.0.*
python-jose[cryptography]==3.3.*
passlib[bcrypt]==1.7.*
google-generativeai==0.8.*
pandas==2.2.*
openpyxl==3.1.*
jinja2==3.1.*
httpx==0.27.*
aiofiles==24.*
python-dateutil==2.9.*
celery==5.4.*
redis==5.*
